+++ duilib/RenderSkia/Render_Skia.h
@@ -32,8 +32,8 @@ public:
 
 	virtual int	GetWidth() override;
 	virtual int GetHeight() override;
-	virtual void ClearAlpha(const UiRect& rcDirty, int alpha = 0) override;
-	virtual void RestoreAlpha(const UiRect& rcDirty, const UiRect& rcShadowPadding, int alpha) override;
+	virtual void ClearAlpha(const UiRect& rcDirty, uint8_t alpha = 0) override;
+	virtual void RestoreAlpha(const UiRect& rcDirty, const UiRect& rcShadowPadding, uint8_t alpha) override;
 	virtual void RestoreAlpha(const UiRect& rcDirty, const UiRect& rcShadowPadding = UiRect()) override;
 
 	virtual bool IsRenderTransparent() const override;
